fix(core): 1.9.1 hotfixes from the open-core readiness audit
- cmd-quality and cmd-bench no longer import the test suite: the case
  corpus moves into redcon/cmd/quality_cases.py so both commands work
  from an installed wheel (they crashed with ModuleNotFoundError for
  every pip user).
- redcon init writes a working docs URL into generated redcon.toml
  (the old one pointed at a repo that does not exist).
- The agents to runtime re-export is lazy now, breaking the import
  cycle that made 'import redcon.runtime.runtime' fail when agents
  loaded first.
- Telemetry sink names are consistent: validation, builder, init
  template and docs all accept noop/file/jsonl/jsonl_file/cloud/http,
  and jsonl_file (the only value that used to pass validation) actually
  builds a file sink instead of silently doing nothing. The telemetry
  doc now tells the truth about the opt-in HttpTelemetrySink and drops
  a link to a page that never existed.
- doctor checks the mcp, symbols and ast_grep extras and whether the
  MCP server is actually registered for any detected agent - the most
  common silent failure was invisible to it.
- pack warns on stderr about an empty result in json format too, and
  points at redcon[tokenizers] when token counts fall back to the
  heuristic because tiktoken is missing.
- License hygiene: LICENSE-COMMERCIAL now covers redcon-cloud/ (it was
  MIT by omission), redcon-cloud/ carries its own license marker, and
  the copyright holder is the real name in both license files.
- Pre-existing ruff debt cleared in every touched file since CI lints
  changed files in full.

Version bumped to 1.9.1.
