+++ LICENSE-COMMERCIAL
@@ -1,9 +1,10 @@
 Redcon Commercial License
-Copyright (c) 2026 nai. All rights reserved.
+Copyright (c) 2026 Natalia Szczepanik. All rights reserved.
 
 The following components of this repository are NOT covered by the MIT License
 and are proprietary to Redcon:
 
+  redcon-cloud/
   redcon/gateway/
   redcon/control_plane/
   redcon/agents/
