+++ redcon-cloud/LICENSE
@@ -1,0 +1,11 @@
+Redcon Commercial License
+
+Copyright (c) 2026 Natalia Szczepanik. All rights reserved.
+
+This directory (redcon-cloud/) is NOT covered by the MIT License in the
+repository root. It is proprietary to Redcon and may not be used, copied,
+modified, merged, published, distributed, sublicensed, or sold without a
+valid commercial license from Redcon.
+
+See LICENSE-COMMERCIAL in the repository root for the full terms.
+For commercial licensing, contact: natjiks@gmail.com
