+++ redcon/cli.py
@@ -677,12 +677,17 @@ def cmd_pack(args: argparse.Namespace) -> int:
         },
     )
 
+    if not data.get("files_included"):
+        # Warn on stderr in every output mode; a silent empty run.json is
+        # indistinguishable from success for a first-time user.
+        print(
+            "warning: no files matched the scan criteria - context is empty",
+            file=sys.stderr,
+        )
+
     if fmt == "json":
         print(_json_mod.dumps(data, indent=2, default=str))
         return 0
-
-    if not data.get("files_included"):
-        print("warning: no files matched the scan criteria - context is empty", file=sys.stderr)
 
     budget = data["budget"]
     cache_report = data.get("cache", {})
@@ -743,6 +748,12 @@ def cmd_pack(args: argparse.Namespace) -> int:
         reason = str(estimator.get("fallback_reason", "") or "")
         if reason:
             _qprint(args, f"Token estimator note: {reason}")
+            if "tiktoken" in reason:
+                _qprint(
+                    args,
+                    "Token estimator hint: install with pip install 'redcon[tokenizers]' "
+                    "for exact counts",
+                )
     summarizer = data.get("summarizer", {})
     if isinstance(summarizer, dict):
         _qprint(
@@ -1676,7 +1687,7 @@ def cmd_init(args: argparse.Namespace) -> int:
     # ── Generate redcon.toml ───────────────────────────────────────────
     lang_comment = f"# Detected dominant language: {dominant_lang}\n" if dominant_lang else ""
     config_content = f"""# Redcon configuration - generated by `redcon init`
-# See: https://github.com/redcon/redcon/docs/configuration.md
+# See: https://github.com/natiixnt/redcon/blob/main/docs/configuration.md
 {lang_comment}
 [budget]
 max_tokens = {default_max_tokens}
@@ -2452,7 +2463,7 @@ def cmd_repo_map(args: argparse.Namespace) -> int:
 def cmd_cmd_quality(args: argparse.Namespace) -> int:
     """Run the M8 quality harness; non-zero exit on any failure (for CI)."""
     from redcon.cmd.quality import run_quality_check
-    from tests.test_cmd_quality import CASES  # type: ignore
+    from redcon.cmd.quality_cases import CASES
 
     failures: list[str] = []
     passed = 0
